:recycle: Move go executables to the apps directory
