+++ apps/selflow-cli/daemon_test.go
