+++ apps/selflow-daemon/project.json
@@ -2,14 +2,14 @@
   "name": "selflow-daemon",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "projectType": "application",
-  "sourceRoot": "cmd/selflow-daemon",
+  "sourceRoot": "",
   "targets": {
     "recreate-daemon": {
       "executor": "nx:run-commands",
       "options": {
         "commands": [
           "mkdir -p tmp",
-          "SELFLOW_DAEMON_HOST_BASED_DIRECTORY=\"$PWD/tmp\" go run github.com/selflow/selflow/cmd/selflow-cli recreate-daemon"
+          "SELFLOW_DAEMON_HOST_BASED_DIRECTORY=\"$PWD/tmp\" go run github.com/selflow/selflow/apps/selflow-cli recreate-daemon"
         ],
         "parallel": false
       },
@@ -19,13 +19,13 @@
       "executor": "@nx-go/nx-go:build",
       "options": {
         "outputPath": "dist/selflow-daemon",
-        "main": "./cmd/selflow-daemon"
+        "main": "./apps/selflow-daemon"
       }
     },
     "serve": {
       "executor": "@nx-go/nx-go:serve",
       "options": {
-        "main": "cmd/selflow-daemon/main.go"
+        "main": "apps/selflow-daemon/main.go"
       }
     },
     "test": {
@@ -37,7 +37,7 @@
     "docker": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "docker build -f ./cmd/selflow-daemon/Dockerfile -t selflow-daemon ."
+        "command": "docker build -f ./apps/selflow-daemon/Dockerfile -t selflow-daemon ."
       }
     }
   },
