+++ apps/selflow-documentation/docs/ecosystem/cli.mdx
@@ -13,7 +13,7 @@ Since Selflow is not ready for production yet, you will need to build the CLI yo
 **Don't worry**, it only need to have [Go](https://go.dev/doc/install) setup on your machine in version `1.20` or up and run this command :
 
 ```bash
-go install github.com/selflow/selflow/cmd/selflow-cli
+go install github.com/selflow/selflow/apps/selflow-cli
 ```
 
 The executable will be added into your `$GOBIN` location. Make sure to add it in your `$PATH`.
