+++ apps/selflow-e2e/src/setup.ts
@@ -1,5 +1,5 @@
 import { execSync } from 'child_process';
 
 export function setup() {
-  //execSync(`go run github.com/selflow/selflow/cmd/selflow-cli recreate-daemon`);
+  //execSync(`go run github.com/selflow/selflow/apps/selflow-cli recreate-daemon`);
 }
