+++ apps/selflow-e2e/src/tools/run.ts
@@ -3,7 +3,7 @@ import * as process from 'process';
 
 export async function startRun(configFilePath: string): Promise<string> {
   const result = execSync(
-    `go run github.com/selflow/selflow/cmd/selflow-cli run \"${configFilePath}\"`,
+    `go run github.com/selflow/selflow/apps/selflow-cli run \"${configFilePath}\"`,
     {
       env: {
         ...process.env,
